fixed device metadata bugs

+++ lambdas/device_api/devices/devices_post.py
@@ -1,6 +1,5 @@
 import json
 from datetime import datetime
-from decimal import Decimal
 
 from boto3.dynamodb.conditions import Key
 
@@ -79,7 +78,7 @@ def upload_device_info(event, context):
             }
         device_id = response[0].get("id")
 
-        date_manufactured = Decimal(str(datetime.fromisoformat(date_manufactured).timestamp()))
+        date_manufactured = int(str(datetime.fromisoformat(date_manufactured).timestamp()))
         print(
             f"Attempting to update device data- firmware_version [{firmware_version}], date_manufactured [{date_manufactured}], notes [{notes}]")
 
@@ -98,7 +97,7 @@ def upload_device_info(event, context):
 
         device_log_table.put_item(Item={
             "device_id": int(device_id),
-            "time": Decimal(str(datetime.now().timestamp())),
+            "time": int(str(datetime.now().timestamp())),
             "log_type": "device_info_connectivity_test",
             "firmware_version": firmware_version,
         })
